feat(foundation-demo): add button-set component and implement in footer
add button-set component and implement in footer; delete daff-footer component
