+++ apps/foundation-demo/src/app/design/molecules/button-set/button-set.md
@@ -1,0 +1,9 @@
+# Button Set Component
+ ## Usage 
+ ```
+<div daff-button-set>
+    <button></button>
+    <button></button>
+    <button></button>
+</daff-button-set>
+```
