+++ apps/foundation-demo/src/app/design/molecules/button-set/button-set.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+
+import { DaffButtonSetComponent } from './components/button-set/button-set.component';
+
+@NgModule({
+  declarations: [
+    DaffButtonSetComponent
+  ],
+  exports: [
+    DaffButtonSetComponent
+  ]
+})
+export class DaffButtonSetModule { }
