+++ apps/foundation-demo/src/app/design/molecules/button-set/components/button-set/button-set.component.scss
@@ -1,0 +1,5 @@
+.daff-button-set {
+  & button {
+    padding: 0 5px;
+  }
+}
