--- apps/foundation-demo/src/app/design/molecules/footer/components/footer-icons/footer-icons.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: '[daff-footer-icons]',
-  template: '<ng-content></ng-content>',
-  host: {'class': 'daff-footer__icons'}
-})
-export class DaffFooterIconsComponent {}
