--- apps/foundation-demo/src/app/design/molecules/footer/components/footer/footer.component.scss
@@ -1,37 +1,0 @@
-@import 'daffodil-util';
-
-.daff-footer {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr) 0.5fr;
-  grid-template-rows: auto;
-  grid-row-gap: 20px;
-  margin: 35px 15px;
-
-  @include breakpoint(medium) {
-    margin: 0 50px 100px 50px;
-    grid-template-columns: repeat(2, 1fr) repeat(2, 0.75fr);
-  }
-
-  @include breakpoint(large) {
-    margin: 0 150px 100px 200px;
-    grid-template-columns: repeat(4, 1fr);
-  }
-
-  &__icons {
-    align-self: end;
-    display: grid;
-    grid-column-gap: 20px;
-    grid-row-gap: 20px;
-    grid-template-columns: repeat(auto-fill, minmax(15px, 1fr));
-    grid-column: 1/-1;
-
-    & > * {
-      margin: auto;
-    }
-
-    @include breakpoint(medium) {
-      grid-column-start: auto;
-      align-self: start;
-    }
-  }
-}
