--- apps/foundation-demo/src/app/design/molecules/footer/footer.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { DaffFooterComponent } from './components/footer/footer.component';
-import { DaffFooterIconsComponent } from './components/footer-icons/footer-icons.component';
-
-@NgModule({
-  declarations: [
-    DaffFooterComponent,
-    DaffFooterIconsComponent,
-  ],
-  exports: [
-    DaffFooterComponent,
-    DaffFooterIconsComponent
-  ]
-})
-export class DaffFooterModule { }
