+++ apps/foundation-demo/src/app/design/molecules/list/list.md
@@ -1,0 +1,10 @@
+# List Component
+ ## Usage 
+ ```
+<ul daff-list>
+  <div daff-list-title>title text</div>
+  <li daff-list-item>list item text 1</li>
+  <li daff-list-item>list item text 2</li>
+  <li daff-list-item>list item text 3</li>
+</ul>
+```
