+++ apps/foundation-demo/src/app/misc/footer/footer.component.html
@@ -1,4 +1,4 @@
-<div daff-footer>
+<div class="footer">
   <ul daff-list>
     <div daff-list-title>support</div>
     <li daff-list-item>Help</li>
@@ -20,9 +20,9 @@
     <li daff-list-item>Women's</li>
     <li daff-list-item>Accessories</li>
   </ul>
-  <div daff-footer-icons>
-    <i class="fab fa-twitter"></i>
-    <i class="fab fa-facebook-f"></i>
-    <i class="fab fa-instagram"></i>
+  <div daff-button-set>
+    <button><i class="fab fa-twitter"></i></button>
+    <button><i class="fab fa-facebook-f"></i></button>
+    <button><i class="fab fa-instagram"></i></button>
   </div>
 </div>
