+++ apps/foundation-demo/src/app/misc/footer/footer.component.scss
@@ -1,0 +1,19 @@
+@import 'daffodil-util';
+
+.footer {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr) 0.6fr;
+  grid-template-rows: auto;
+  grid-row-gap: 20px;
+  margin: 35px 15px;
+
+  @include breakpoint(medium) {
+    margin: 0 50px 100px 50px;
+    grid-template-columns: repeat(2, 1fr) repeat(2, 0.75fr);
+  }
+
+  @include breakpoint(large) {
+    margin: 0 150px 100px 200px;
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
