+++ apps/foundation-demo/src/app/misc/misc.module.ts
@@ -4,14 +4,14 @@ import { CommonModule } from '@angular/common';
 import { NotFoundComponent } from './not-found/not-found.component';
 import { HelpBoxComponent } from './help-box/help-box.component';
 import { FooterComponent } from './footer/footer.component';
-import { DaffFooterModule } from '../design/molecules/footer/footer.module';
 import { DaffListModule } from '../design/molecules/list/list.module';
+import { DaffButtonSetModule } from '../design/molecules/button-set/button-set.module';
 
 @NgModule({
   imports: [
     CommonModule,
-    DaffFooterModule,
-    DaffListModule
+    DaffListModule,
+    DaffButtonSetModule
   ],
   declarations: [
     NotFoundComponent,
